+++ src/logger/tests/LoggerFactoryTest.php
@@ -16,6 +16,10 @@ use Hyperf\Config\Config;
 use Hyperf\Contract\ConfigInterface;
 use Hyperf\Logger\LoggerFactory;
 use Hyperf\Utils\ApplicationContext;
+use Hyperf\Utils\Context;
+use HyperfTest\Logger\Stub\BarProcessor;
+use HyperfTest\Logger\Stub\FooHandler;
+use HyperfTest\Logger\Stub\FooProcessor;
 use Mockery;
 use Monolog\Handler\StreamHandler;
 use Monolog\Handler\TestHandler;
@@ -32,6 +36,7 @@ class LoggerFactoryTest extends TestCase
     public function tearDown()
     {
         Mockery::close();
+        Context::set('test.logger.foo_handler.record', null);
     }
 
     public function testInvokeLoggerFactory()
@@ -103,6 +108,60 @@ class LoggerFactoryTest extends TestCase
         $this->assertInstanceOf(TestHandler::class, $handlers[1]);
     }
 
+    public function testNotSetProcessor()
+    {
+        $container = $this->mockContainer();
+        $factory = $container->get(LoggerFactory::class);
+        $logger = $factory->get('hyperf');
+        $reflectionClass = new ReflectionClass($logger);
+        $handlersProperty = $reflectionClass->getProperty('processors');
+        $handlersProperty->setAccessible(true);
+        $processors = $handlersProperty->getValue($logger);
+        $this->assertSame([], $processors);
+    }
+
+    public function testProcessor()
+    {
+        $container = $this->mockContainer();
+        $factory = $container->get(LoggerFactory::class);
+        $logger = $factory->get('hyperf', 'processor-test');
+        $reflectionClass = new ReflectionClass($logger);
+        $handlersProperty = $reflectionClass->getProperty('processors');
+        $handlersProperty->setAccessible(true);
+        $processors = $handlersProperty->getValue($logger);
+        $this->assertSame(3, count($processors));
+        $this->assertInstanceOf(FooProcessor::class, $processors[0]);
+
+        $logger->info('Hello world.');
+
+        $this->assertSame(
+            'Hello world.Hello world.',
+            Context::get('test.logger.foo_handler.record')['message']
+        );
+        $this->assertTrue(Context::get('test.logger.foo_handler.record')['bar']);
+        $this->assertTrue(Context::get('test.logger.foo_handler.record')['callback']);
+    }
+
+    public function testDefaultProcessor()
+    {
+        $container = $this->mockContainer();
+        $factory = $container->get(LoggerFactory::class);
+        $logger = $factory->get('hyperf', 'default-processor');
+        $reflectionClass = new ReflectionClass($logger);
+        $handlersProperty = $reflectionClass->getProperty('processors');
+        $handlersProperty->setAccessible(true);
+        $processors = $handlersProperty->getValue($logger);
+        $this->assertSame(1, count($processors));
+        $this->assertInstanceOf(FooProcessor::class, $processors[0]);
+
+        $logger->info('Hello world.');
+
+        $this->assertSame(
+            'Hello world.Hello world.',
+            Context::get('test.logger.foo_handler.record')['message']
+        );
+    }
+
     private function mockContainer(): ContainerInterface
     {
         $container = Mockery::mock(ContainerInterface::class);
@@ -147,6 +206,55 @@ class LoggerFactoryTest extends TestCase
                     'formatter' => [
                         'class' => \Monolog\Formatter\LineFormatter::class,
                         'constructor' => [],
+                    ],
+                ],
+                'processor-test' => [
+                    'handlers' => [
+                        [
+                            'class' => FooHandler::class,
+                            'constructor' => [
+                                'stream' => BASE_PATH . '/runtime/logs/hyperf.log',
+                                'level' => \Monolog\Logger::DEBUG,
+                            ],
+                            'formatter' => [
+                                'class' => \Monolog\Formatter\LineFormatter::class,
+                            ],
+                        ],
+                    ],
+                    'processors' => [
+                        [
+                            'class' => FooProcessor::class,
+                            'constructor' => [
+                                'repeat' => 2,
+                            ],
+                        ],
+                        [
+                            'class' => BarProcessor::class,
+                        ],
+                        function (array $records) {
+                            $records['callback'] = true;
+                            return $records;
+                        },
+                    ],
+                ],
+                'default-processor' => [
+                    'handlers' => [
+                        [
+                            'class' => FooHandler::class,
+                            'constructor' => [
+                                'stream' => BASE_PATH . '/runtime/logs/hyperf.log',
+                                'level' => \Monolog\Logger::DEBUG,
+                            ],
+                            'formatter' => [
+                                'class' => \Monolog\Formatter\LineFormatter::class,
+                            ],
+                        ],
+                    ],
+                    'processor' => [
+                        'class' => FooProcessor::class,
+                        'constructor' => [
+                            'repeat' => 2,
+                        ],
                     ],
                 ],
             ],